+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,11 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Added
 
 - `limestone/Card` `splitCaption` prop to allow user to split the content for the captions
+- `ui/useScroll.ScrollAnimator` `oase-out-cubic` animation function
+
+### Changed
+
+- `ui/useScroll.ScrollAnimator` default animation function
 
 ## [5.3.2] - 2025-10-28